chat-service/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.1
openai==1.88.0
pymongo==4.7.3
pytest==8.3.5
Requests==2.32.4
google-genai>=1.50.1
